+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## vx.x.x
 * Fix vtk version #291
+* Disables buttons in Select-Image tab after first registration #293
 * Edit DVC-Results tab #231
 * Edit mask/point-cloud/results tooltips, help text and documentation #257 #286 #287
 * Scales the displacement vectors keeping the color bar with the displacement values. Adds title to color bar #270